+++ module/constants.js
@@ -9,5 +9,6 @@ export const Constants = {
         colors_value: "colors_value",
         colors_hue: "colors_hue",
         colors_lightness: "colors_lightness",
+        compact_mode: "compact_mode"
     }
 }